engine: warnings are errors, and there are none
poly_engine's 53 warning sites are gone: int indexes into std::array gain
static_cast<size_t>, the float-to-double promotions in rng.h and engine.cpp
are spelled out, and the int64-to-double and int-to-float sites cast to the
type the expression already produced. No arithmetic changed type, and the
701 unit and golden tests are unchanged. GCC 15's -Wfree-nonheap-object from
inside libstdc++'s vector reallocation, reached from smf_writer.cpp, is a
false positive and is suppressed for GCC only, file-wide, with the reason.

POLY_ENGINE_WARNINGS_FATAL, ON by default, makes the engine -Werror/-WX on
every build; tests, tools and the plugin still follow POLY_WARNINGS_FATAL,
OFF until their own burndown, which starts at 63 sites recorded in the
decisions file. The engine-isolation job passes the option explicitly.
Proved clean on GCC 15.2, Apple Clang 21 and Emscripten; a planted sign
conversion failed GCC and Clang before it was removed.

M002/S03 closes, and with it M002: every workflow declares its grant, the
untrusted inputs are fuzzed, and an engine warning fails the build.

Docs-Not-Affected: docs/engine-spec.md — every engine change is an explicit cast spelling out a conversion the compiler already performed; no value, range or behaviour the spec describes moved, and unit's golden tests are unchanged

Plan: docs/plans/open-source-launch/ledger.md
Slice: M002/S03
Rows: OS11

+++ engine/include/poly/rng.h
@@ -51,7 +51,7 @@ inline float correlatedNoise(uint64_t seed, int laneId, int64_t absStep) {
         const float a = deterministicRand(seed, laneId, cell, channel) * 2.0f - 1.0f;
         const float b = deterministicRand(seed, laneId, cell + 1, channel) * 2.0f - 1.0f;
         const double smooth = frac * frac * (3.0 - 2.0 * frac);
-        return static_cast<float>(a + (b - a) * smooth);
+        return static_cast<float>(static_cast<double>(a) + static_cast<double>(b - a) * smooth);
     };
 
     // Halving amplitude over lengthening periods: the 1/f shape. Weights sum to
